+++ changelog.md
@@ -2,11 +2,18 @@
 
 ## Version 0.33.0
 Happy new year!
-# New Features
+### New Features
 1. You can now change the color of friends
 1. Error toast when the queue mess up (you get disconnected)
-1. Added a few simple api's (openPacks, eventManager)
-# Fixes
+1. Added a few simple APIs
+  <extended>
+    `underscript.`
+    1. openPacks
+    1. eventManager
+  </extended>
+### Improvements
+1. Improved pack opening (less likely to crash with over 2000 packs)
+### Fixes
 1. Fixed server-restart not being detected
 1. Fixed removing cosmetics highlight in the menu